dev
A very basic broken dashboard that is best used as a boilerplate. Right now the only thing it does decent is graph out commodities. We need to abstract CPI from inflation, and properly track the gdp of several countries. Additionally, we need to accuratly track all of the interest rates for the 5 currencies we are interested in